Added product reserving in cart purchase
When purchasing cart, we now reserve all items, try to use external services, and if they fail we "rollback" the reservations.
If all went good, we request the transaction to be committed.
No need to lock anything but a single product each time we reserve a certain amount from it.
NEED TO PASS A FEW MORE TESTS